structuring file
